refactor: 현재 인증된 사용자 정보를 String으로 조회할 수 있도록 변경
- 기존 User 객체 등록에서 사용자의 googleId로 조회할 수 있다.

+++ server/src/main/java/wap/starlist/bookmark/domain/Bookmark.java
@@ -40,7 +40,7 @@ public class Bookmark {
 
     private Boolean syncing;
 
-    private Integer googleId;
+    private Long googleId;
 
     @ManyToOne
     private Folder folderId;
